+++ src/utils/readAsDataUrl.ts
@@ -1,0 +1,10 @@
+export function readAsDataUrl(file: Blob): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+
+    reader.addEventListener('load', () => resolve(reader.result));
+    reader.addEventListener('error', () => reject(new Error('Unable to read file')));
+
+    reader.readAsDataURL(file);
+  });
+}