+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased]
+## [1.7.0] - 2017-12-03
 ### Added
 - New Quick Links plugin for your favourite websites. Try using the keyboard shortcut!